[Resolves #159952028] Fix registration workflow
# Story Title

[Fix registration workflow](https://www.pivotaltracker.com/story/show/159952028)

# Changes made

- Add `role` params to existing sign in route. Depending on which carousel slide user navigated to the sign in page from, role will get chosen by default.
- Add `vue-form-wizard` library to project
- Granulate bulk of `SignupPage/index` code to `SignupPage/SignupTemplate` component as this component needs to be reused.
- Granulate css on `SignupPage/index`.
- Create separate `RegisterClub` component to for users signing up as director.
- Write update operation, route and contracts for `Club`
- Redirect use to `RegisterClub` component instead of sign in page if user tried to register as a director after signing in. 

# How does the solution addresses the problem

Logically speaking, a single club cannot have multiple users with the director(aka account owner) role. Therefore, we need to fix the registration workflow such that if a user tried to register as a director, we need to check if the club already has an existing director or not. This is being done so by using the `vue-form-wizard` library to create a registration work flow with multiple steps. After registering as a user with the role of director, the user will be brought to a second component that prompts them to choose their club, after choosing their club, an API endpoint will be called to check if that club already has an existing owner, if yes, they cannot register for that club. If not, the club will be chosen and if the user clicks on the register club button, it will call the newly created API endpoint to update the owner of the club.