[Resolves #151945495] Viewing notes FE
# Story Title

[Viewing notes FE](https://www.pivotaltracker.com/story/show/151945495)

# Changes made

- Added author name and updated at to note representer
- Created NoteService to fetch index and create notes on the server side
- Updated UI to use data fetched from the server
- Created namespaced store for notes under club
- Uniformed button styling

# How does the solution addresses the problem

This allows the clubs to load their notes and view their contents.
At the moment, notes being loaded are the user notes and not the club specific notes as I need to think about the best way to distinguish them and how to fetch them. Currently implemented solution for posts seems a bit overkill in terms of duplication.