[Chore] Fix jobs list
# Story Title

Fix jobs list

# Changes made

- Commented my assignments as this needs to move to a separate page
- Fixed displaying of the my reports
- Fixed navbar to have the menu points for scouting properly defined
- Removed usage of search.scss and wrote simplified css for the needed style for table header
- Created report row component and using that report row for my reports table instead of the massively complex report item component previously created.

# How does the solution addresses the problem

After applying this, my reports list is displayed properly, using a component specific for this row. This can probably be further cleanup and reused for other report listing but along the way, with the usage we'll see what can we do to cleanup this further. The key thing for this patch is to fix the displaying of my reports and the update of the navbar for the scouts.