[Chore] Fix search errors
# Story Title

Fix search errors

# Changes made

- Using representer for index operations when searching for a competition or a club
- Updated the client side to expect the key `teams` or `competitions` when receiving the results from the server

# How does the solution addresses the problem

Before this being merged, the results of the search operation would vary if they had results to return or not. This was causing the select to throw an error due to the unexpected format. Using the representer ensures that the format sent to the client side is always the same and the client side can parse the results properly.