[Hotfix] Fix profile picture update on production env
# Story Title

N/a

# Changes made

- Updated nginx configuration to pass rails active storage routes to puma
- Set Sidekiq as Queue Processor
- Set the default url to 'firsttouch.io'

# How does the solution addresses the problem

Before merging this in, because the default url was not set, trying to fetch the url for the profile image was failing and causing the render of the user to crash.
Another problem was caused by nginx not redirecting `rails/active_storage` to puma, when trying to load the image. So, before merging this one in the user couldn't fetch its profile image.